Remove obsolete confirmation workflow and add Sim Head Unit service
