+++ SimHeadUnit/requirements.txt
@@ -1,0 +1,4 @@
+fastapi==0.115.5
+uvicorn==0.32.0
+ankaios_sdk
+requests
